Refactor event management localization: removed event locales.json and updated imports to admin/locales.json across components. Deleted AdminEventPage and its associated logic for admin event creation.
