+++ src/components/admin/event/OrganizerSection.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 
 import { SelectAlt, SelectContentAlt, SelectItemAlt, SelectTriggerAlt, SelectValueAlt } from '@/components/ui/selectAlt'
 import { getAllOrganizersAction } from '@/app/[locale]/admin/actions'
-import Translations from '@/app/[locale]/event/locales.json'
+import Translations from '@/app/[locale]/admin/locales.json'
 import { getTranslations } from '@/lib/getDictionary'
 import { Organizer } from '@/models/organizer.model'
 import { Label } from '@/components/ui/label'
